move fuse support to the hardware independent section
