+++ src/fs/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "pci")]
+pub mod fuse;
+
+pub fn init() {
+	#[cfg(feature = "pci")]
+	fuse::init();
+}
